'end_with' nodes without prerequisites are no real end nodes
So they should be treated like normal nodes, i.e. we
- do not paint them green,
- do not add URLs to them,
- do not call end node callbacks for them which might otherwise cause the
  creation of trivial, one-node sub-graphs,

Please note that there also is one incompatible little change in behaviour:
'end_with' nodes, no matter if with or without prerequisites, now get their
commands/recipes printed because this is what a user would expect for a node
occurring in a graph with 'trim_mode == false'. One test file was changed to
reflect this new behaviour.